# nest-template/openapi-schema.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/users": {
            "post": {
                "operationId": "UsersController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUsers"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "UsersController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadUsersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "operationId": "UsersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUsers"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "UsersController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "UsersController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneUsersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{id}/soft": {
            "delete": {
                "operationId": "UsersController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/sellers": {
            "post": {
                "operationId": "SellersController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSellers"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "SellersController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadSellersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/sellers/{id}": {
            "put": {
                "operationId": "SellersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSellers"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "SellersController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "SellersController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneSellersQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/sellers/{id}/soft": {
            "delete": {
                "operationId": "SellersController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/reviews": {
            "post": {
                "operationId": "ReviewsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviews"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "ReviewsController_read",
                "parameters": [
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadReviewsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/reviews/{id}": {
            "put": {
                "operationId": "ReviewsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReviews"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "ReviewsController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "get": {
                "operationId": "ReviewsController_readOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ReadOneReviewsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/reviews/{id}/soft": {
            "delete": {
                "operationId": "ReviewsController_softDelete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "info": {
        "title": "Nestjs Template API",
        "description": "",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateReviews": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "product_id": {
                        "nullable": true
                    },
                    "rating": {
                        "type": "number"
                    },
                    "review_text": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "rating"
                ],
                "additionalProperties": false
            },
...
            "ReadReviewsQuery": {
                "type": "object",
                "properties": {
                    "filters": {
                        "$ref": "#/components/schemas/ReadReviewsFilters"
                    },
                    "relations": {
                        "type": "object",
                        "properties": {
                            "user_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadUsersRelations"
                                    }
                                ]
                            },
                            "product_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadProductsRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    },
                    "orders": {
                        "$ref": "#/components/schemas/ReadReviewsOrders"
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "skip": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 0
                            },
                            "take": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 25
                            }
                        },
                        "required": [],
                        "additionalProperties": false,
                        "default": {
                            "skip": 0,
                            "take": 25
                        }
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "UpdateReviews": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "product_id": {
                        "nullable": true
                    },
                    "rating": {
                        "type": "number"
                    },
                    "review_text": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "CreateSellers": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "store_name": {
                        "type": "string",
                        "maxLength": 100
                    }
                },
                "required": [
                    "store_name"
                ],
                "additionalProperties": false
            },
            "ReadOneSellersQuery": {
                "type": "object",
                "properties": {
                    "relations": {
                        "type": "object",
                        "properties": {
                            "user_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadUsersRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "ReadSellersOrders": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "anyOf": [
                            {
                                "enum": [
                                    "asc",
                                    "desc"
                                ]
                            },
                            {
                                "$ref": "#/components/schemas/ReadUsersOrders"
                            }
                        ]
                    },
                    "store_name": {
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "ReadSellersQuery": {
                "type": "object",
                "properties": {
                    "filters": {
                        "$ref": "#/components/schemas/ReadSellersFilters"
                    },
                    "relations": {
                        "type": "object",
                        "properties": {
                            "user_id": {
                                "anyOf": [
                                    {
                                        "anyOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "boolean"
                                            }
                                        ],
                                        "type": "boolean"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ReadUsersRelations"
                                    }
                                ]
                            }
                        },
                        "required": [],
                        "additionalProperties": false
                    },
                    "orders": {
                        "$ref": "#/components/schemas/ReadSellersOrders"
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "skip": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 0
                            },
                            "take": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 25
                            }
                        },
                        "required": [],
                        "additionalProperties": false,
                        "default": {
                            "skip": 0,
                            "take": 25
                        }
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "UpdateSellers": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "nullable": true
                    },
                    "store_name": {
                        "type": "string",
                        "maxLength": 100
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "CreateUsers": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "password_hash": {
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "required": [
                    "username",
                    "email",
                    "password_hash"
                ],
                "additionalProperties": false
            },
            "ReadOneUsersQuery": {
                "type": "object",
                "properties": {
                    "relations": {
                        "type": "object",
                        "properties": {},
                        "required": [],
                        "additionalProperties": false
                    }
                },
                "required": [],
                "additionalProperties": false
            },
...
            "ReadUsersOrders": {
                "type": "object",
                "properties": {
                    "username": {
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "email": {
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "password_hash": {
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "ReadUsersQuery": {
                "type": "object",
                "properties": {
                    "filters": {
                        "$ref": "#/components/schemas/ReadUsersFilters"
                    },
                    "relations": {
                        "type": "object",
                        "properties": {},
                        "required": [],
                        "additionalProperties": false
                    },
                    "orders": {
                        "$ref": "#/components/schemas/ReadUsersOrders"
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "skip": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 0
                            },
                            "take": {
                                "anyOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "number"
                                    }
                                ],
                                "type": "number",
                                "default": 25
                            }
                        },
                        "required": [],
                        "additionalProperties": false,
                        "default": {
                            "skip": 0,
                            "take": 25
                        }
                    }
                },
                "required": [],
                "additionalProperties": false
            },
            "ReadUsersRelations": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
            },
            "UpdateUsers": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "password_hash": {
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "required": [],
                "additionalProperties": false
            }
        }
    },
...
}